Limit background notifications to mobile targets
