+++ composeApp/src/desktopMain/kotlin/com/hansholz/bestenotenapp/main.kt
@@ -27,7 +27,6 @@ import com.hansholz.bestenotenapp.main.ExactPlatform
 import com.hansholz.bestenotenapp.main.LocalNavigationDrawerTopPadding
 import com.hansholz.bestenotenapp.main.LocalTitleBarModifier
 import com.hansholz.bestenotenapp.main.getExactPlatform
-import com.hansholz.bestenotenapp.notifications.ensureDesktopNotificationsInitialized
 import com.jetbrains.JBR
 import java.awt.Color
 import java.awt.Dimension
@@ -36,7 +35,6 @@ import org.jetbrains.skiko.hostOs
 
 @OptIn(ExperimentalMaterial3ComponentOverrideApi::class, ExperimentalMaterial3Api::class)
 fun main() {
-    ensureDesktopNotificationsInitialized()
     application {
         DecoratedWindow(
             onCloseRequest = ::exitApplication,
